+++ actioncable/test/channel/stream_test.rb
@@ -104,7 +104,24 @@ module ActionCable::StreamTests
 
         pubsub_call = channel.pubsub.class.class_variable_get "@@subscribe_called"
 
-        assert_equal "action_cable:stream_tests:chat:Room#1-Campfire", pubsub_call[:channel]
+        assert_equal "action_cable:stream_tests:chat:Room#1Campfire", pubsub_call[:channel]
+        assert_instance_of Proc, pubsub_call[:callback]
+        assert_instance_of Proc, pubsub_call[:success_callback]
+      end
+    end
+
+    test "stream_for_list" do
+      run_in_eventmachine do
+        connection = TestConnection.new
+
+        channel = ChatChannel.new connection, ""
+        channel.subscribe_to_channel
+        channel.stream_for_list [Room.new(1), Room.new(2)]
+        wait_for_async
+
+        pubsub_call = channel.pubsub.class.class_variable_get "@@subscribe_called"
+
+        assert_equal "action_cable:stream_tests:chat:Room#1Campfire-Room#2Campfire", pubsub_call[:channel]
         assert_instance_of Proc, pubsub_call[:callback]
         assert_instance_of Proc, pubsub_call[:success_callback]
       end
@@ -121,7 +138,7 @@ module ActionCable::StreamTests
 
         pubsub_call = channel.pubsub.class.class_variable_get "@@subscribe_called"
 
-        assert_equal "action_cable:stream_tests:chat:Room#1-Campfire", pubsub_call[:channel]
+        assert_equal "action_cable:stream_tests:chat:Room#1Campfire", pubsub_call[:channel]
         assert_instance_of Proc, pubsub_call[:callback]
         assert_instance_of Proc, pubsub_call[:success_callback]
       end